requirements.txt
pyyaml
traits
